python review
